askadmin/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
annotated-types==0.6.0
    # via pydantic
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
    #   django-ninja-extra
cffi==1.16.0
    # via cryptography
contextlib2==21.6.0
    # via django-ninja-extra
cryptography==42.0.5
    # via pyjwt
django==5.0.4
    # via
    #   django-cors-headers
    #   django-ninja
    #   django-ninja-extra
    #   django-ninja-jwt
django-cors-headers==4.3.1
django-ninja==1.1.0
    # via django-ninja-extra
django-ninja-extra==0.20.5
    # via django-ninja-jwt
django-ninja-jwt==5.3.0
injector==0.21.0
    # via django-ninja-extra
pycparser==2.22
    # via cffi
pydantic==2.6.4
    # via django-ninja
pydantic-core==2.16.3
    # via pydantic
pyjwt==2.8.0
    # via django-ninja-jwt
sqlparse==0.4.4
    # via django
typing-extensions==4.10.0
    # via
    #   pydantic
    #   pydantic-core
